fix: questionnaire answers would not be registered.
This due to the CommonObject::createCommon() method, at line 8391:
`$values[$k] = $this->quote($v, $value); // May return string 'NULL' if $value is null`

If fk_choix and fk_choix_col are not defined, they are set to NULL, which is forbidden.
This commit set them to the default value, which is 0.